requirements.txt
docxtpl==0.11.3
pyyaml==5.4.1
